moved detector includes to subfolder
