+++ ctbGui/ctbGui.cpp
@@ -9,7 +9,7 @@
 #include <fstream>
 #include <string>
 
-#include "Detector.h"
+#include "sls/Detector.h"
 #include "sls_detector_defs.h"
 //#include "sls_receiver_defs.h"
 #include "ctbMain.h"
