+++ ctbGui/ctbPowers.cpp
@@ -13,7 +13,7 @@
 #include "ctbDefs.h"
 #include "ctbDacs.h"
 #include "ctbPowers.h"
-#include "Detector.h"
+#include "sls/Detector.h"
 #include "sls_detector_defs.h"
 
 using namespace std;
