+++ ctbGui/ctbSignals.cpp
@@ -35,7 +35,7 @@
 
 #include "ctbSignals.h"
 #include "ctbDefs.h"
-#include "Detector.h"
+#include "sls/Detector.h"
 
 using namespace std;
