+++ ctbGui/ctbSlowAdcs.cpp
@@ -10,7 +10,7 @@
 
 #include "ctbSlowAdcs.h"
 #include "ctbDefs.h"
-#include "Detector.h"
+#include "sls/Detector.h"
 #include "sls_detector_defs.h"
 
 using namespace std;
