+++ docs/src/gendoc.cpp
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include "CmdProxy.h"
-#include "Detector.h"
+#include "sls/Detector.h"
 #include "sls_detector_defs.h"
 
 std::string replace_all(const std::string &src, const std::string &from,
