+++ slsDetectorGui/include/qDacWidget.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Detector.h"
+#include "sls/Detector.h"
 #include "sls_detector_defs.h"
 #include "ui_form_dac.h"
 #include <string>
