+++ slsDetectorSoftware/include/sls/Detector.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Result.h"
+#include "sls/Result.h"
 #include "network_utils.h"
 #include "sls_detector_defs.h"
 #include <chrono>
