+++ slsDetectorSoftware/include/sls/Result.h
