+++ slsDetectorSoftware/include/sls/detectorData.h
