+++ slsDetectorSoftware/tests/test-Result.cpp
@@ -1,4 +1,4 @@
-#include "Result.h"
+#include "sls/Result.h"
 #include "ToString.h"
 #include "TypeTraits.h"
 #include "catch.hpp"
